09.07.2015

+++ src/main/java/com/domain/UserRole.java
@@ -18,7 +18,7 @@ public class UserRole {
 	private String name;
 
     @Column(name="age")
-    private int age;
+    private Integer age;
 
     @OneToOne(cascade = CascadeType.ALL)
 	private User user;
@@ -30,13 +30,13 @@ public class UserRole {
         this.user = user;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
     public String getName() {
         return name;
     }
-    public long getAge() {
+    public Integer getAge() {
         return age;
     }
     public User getUser() {
